+++ api/webhook.js
@@ -1,7 +1,7 @@
 /// @ts-check
-const {createProbot} = require('probot');
-const {resolve} = require('probot/lib/resolver');
-const {findPrivateKey} = require('probot/lib/private-key');
+const {Probot} = require('probot');
+const {resolveAppFunction} = require('probot/lib/helpers/resolve-app-function');
+const {findPrivateKey} = require('probot/lib/helpers/get-private-key');
 
 const githubApp = require('@graphql-inspector/github').app;
 
@@ -10,14 +10,14 @@ let probot;
 const loadProbot = (appFn) => {
   probot =
     probot ||
-    createProbot({
+    new Probot({
       id: process.env.APP_ID,
       secret: process.env.WEBHOOK_SECRET,
       cert: findPrivateKey(),
     });
 
   if (typeof appFn === 'string') {
-    appFn = resolve(appFn);
+    appFn = resolveAppFunction(appFn);
   }
 
   probot.load(appFn);