+++ packages/commands/coverage/src/index.ts
@@ -108,11 +108,20 @@ export default createCommand<
       const writePath = args.write;
       const silent = args.silent;
       const {headers, token} = parseGlobalArgs(args);
+      const apolloFederation = args.federation || false;
+      const aws = args.aws || false;
+      const method = args.method?.toUpperCase() || 'POST';
 
-      const schema = await loaders.loadSchema(args.schema, {
-        token,
-        headers,
-      });
+      const schema = await loaders.loadSchema(
+        args.schema,
+        {
+          token,
+          headers,
+          method,
+        },
+        apolloFederation,
+        aws,
+      );
       const documents = await loaders.loadDocuments(args.documents);
 
       return handler({schema, documents, silent, writePath});