+++ packages/commands/serve/src/index.ts
@@ -7,7 +7,7 @@ import {
 import {Logger} from '@graphql-inspector/logger';
 import open from 'open';
 import express from 'express';
-import graphql from 'express-graphql';
+import {graphqlHTTP} from 'express-graphql';
 import cors from 'cors';
 import {fake} from './fake';
 
@@ -43,10 +43,20 @@ export default createCommand<
     },
     async handler(args) {
       const {headers, token} = parseGlobalArgs(args);
-      const schema = await loaders.loadSchema(args.schema, {
-        headers,
-        token,
-      });
+      const apolloFederation = args.federation || false;
+      const aws = args.aws || false;
+      const method = args.method?.toUpperCase() || 'POST';
+
+      const schema = await loaders.loadSchema(
+        args.schema,
+        {
+          headers,
+          token,
+          method,
+        },
+        apolloFederation,
+        aws,
+      );
 
       const port = args.port;
 
@@ -61,7 +71,7 @@ export default createCommand<
 
         app.use(
           cors(),
-          graphql({
+          graphqlHTTP({
             schema,
             graphiql: true,
           }),