+++ packages/commands/similar/src/index.ts
@@ -126,11 +126,20 @@ export default createCommand<
       const writePath = args.write;
       const type = args.name;
       const threshold = args.threshold;
+      const apolloFederation = args.federation || false;
+      const aws = args.aws || false;
+      const method = args.method?.toUpperCase() || 'POST';
 
-      const schema = await loaders.loadSchema(args.schema, {
-        headers,
-        token,
-      });
+      const schema = await loaders.loadSchema(
+        args.schema,
+        {
+          headers,
+          token,
+          method,
+        },
+        apolloFederation,
+        aws,
+      );
 
       return handler({schema, writePath, type, threshold});
     },