+++ packages/core/src/diff/field.ts
@@ -1,12 +1,10 @@
 import {
   GraphQLField,
   GraphQLObjectType,
-  GraphQLArgument,
   GraphQLInterfaceType,
 } from 'graphql';
 
-import {isNotEqual, isVoid} from './common/compare';
-import {Change} from './changes/change';
+import {isNotEqual, isVoid} from '../utils/compare';
 import {
   fieldDescriptionChanged,
   fieldDescriptionAdded,
@@ -21,71 +19,56 @@ import {
   fieldArgumentRemoved,
 } from './changes/field';
 import {changesInArgument} from './argument';
-import {unionArrays, diffArrays} from '../utils/arrays';
+import {compareLists} from '../utils/compare';
+import { AddChange } from './schema';
 
 export function changesInField(
   type: GraphQLObjectType | GraphQLInterfaceType,
   oldField: GraphQLField<any, any>,
   newField: GraphQLField<any, any>,
-): Change[] {
-  const changes: Change[] = [];
-
+  addChange: AddChange
+) {
   if (isNotEqual(oldField.description, newField.description)) {
     if (isVoid(oldField.description)) {
-      changes.push(fieldDescriptionAdded(type, newField));
+      addChange(fieldDescriptionAdded(type, newField));
     } else if (isVoid(newField.description)) {
-      changes.push(fieldDescriptionRemoved(type, oldField));
+      addChange(fieldDescriptionRemoved(type, oldField));
     } else {
-      changes.push(fieldDescriptionChanged(type, oldField, newField));
+      addChange(fieldDescriptionChanged(type, oldField, newField));
     }
   }
 
   if (isNotEqual(oldField.isDeprecated, newField.isDeprecated)) {
     if (newField.isDeprecated) {
-      changes.push(fieldDeprecationAdded(type, newField));
+      addChange(fieldDeprecationAdded(type, newField));
     } else {
-      changes.push(fieldDeprecationRemoved(type, oldField));
+      addChange(fieldDeprecationRemoved(type, oldField));
     }
   }
 
   if (isNotEqual(oldField.deprecationReason, newField.deprecationReason)) {
     if (isVoid(oldField.deprecationReason)) {
-      changes.push(fieldDeprecationReasonAdded(type, newField));
+      addChange(fieldDeprecationReasonAdded(type, newField));
     } else if (isVoid(newField.deprecationReason)) {
-      changes.push(fieldDeprecationReasonRemoved(type, oldField));
+      addChange(fieldDeprecationReasonRemoved(type, oldField));
     } else {
-      changes.push(fieldDeprecationReasonChanged(type, oldField, newField));
+      addChange(fieldDeprecationReasonChanged(type, oldField, newField));
     }
   }
 
   if (isNotEqual(oldField.type.toString(), newField.type.toString())) {
-    changes.push(fieldTypeChanged(type, oldField, newField));
+    addChange(fieldTypeChanged(type, oldField, newField));
   }
 
-  const oldArgs = oldField.args;
-  const newArgs = newField.args;
-  const oldNames = oldArgs.map((a) => a.name);
-  const newNames = newArgs.map((a) => a.name);
-
-  const added = diffArrays(newNames, oldNames).map(
-    (name) => newArgs.find((a) => a.name === name) as GraphQLArgument,
-  );
-  const removed = diffArrays(oldNames, newNames).map(
-    (name) => oldArgs.find((a) => a.name === name) as GraphQLArgument,
-  );
-  const common = unionArrays(oldNames, newNames).map((name) => ({
-    inOld: oldArgs.find((a) => a.name === name) as GraphQLArgument,
-    inNew: newArgs.find((a) => a.name === name) as GraphQLArgument,
-  }));
-
-  common.forEach(({inOld, inNew}) => {
-    changes.push(...changesInArgument(type, oldField, inOld, inNew));
-  });
-
-  changes.push(...added.map((arg) => fieldArgumentAdded(type, newField, arg)));
-  changes.push(
-    ...removed.map((arg) => fieldArgumentRemoved(type, oldField, arg)),
-  );
-
-  return changes;
+  compareLists(oldField.args, newField.args, {
+    onAdded(arg) {
+      addChange(fieldArgumentAdded(type, newField, arg));
+    },
+    onRemoved(arg) {
+      addChange(fieldArgumentRemoved(type, oldField, arg));
+    },
+    onMutual(arg) {
+      changesInArgument(type, oldField, arg.oldVersion, arg.newVersion, addChange)
+    }
+  })
 }