+++ packages/core/src/diff/rules/suppress-removal-of-deprecated-field.ts
@@ -2,7 +2,7 @@ import {isObjectType, isInterfaceType} from 'graphql';
 
 import {CriticalityLevel, ChangeType} from './../changes/change';
 import {Rule} from './types';
-import {parsePath} from '../common/path';
+import {parsePath} from '../../utils/path';
 
 export const suppressRemovalOfDeprecatedField: Rule = ({
   changes,