+++ packages/github/src/helpers/config.ts
@@ -19,7 +19,7 @@ export interface SchemaPointer {
 export interface LegacyConfig {
   diff?: boolean | Diff;
   notifications?: Notifications;
-  endpoint?: string;
+  endpoint?: Endpoint;
   schema: SchemaPointer;
 }
 
@@ -30,14 +30,17 @@ export interface Notifications {
 }
 
 interface Diff {
+  experimental_merge?: boolean;
   annotations?: boolean;
   failOnBreaking?: boolean;
+  approveLabel?: string;
   intercept?: DiffInterceptor;
+  summaryLimit?: number;
 }
 
 interface Environment {
   branch: string;
-  endpoint?: string;
+  endpoint?: Endpoint;
   diff?: Diff | boolean;
   notifications?: Notifications | boolean;
 }
@@ -46,7 +49,7 @@ export interface NormalizedEnvironment {
   name: string;
   schema: string;
   branch: string;
-  endpoint?: string;
+  endpoint?: Endpoint;
   diff: Diff | false;
   notifications: Notifications | false;
 }
@@ -93,20 +96,28 @@ const diffDefault = {
 };
 const notificationsDefault = false;
 
-function normalizeConfig(config: Config): NormalizedConfig {
+function normalizeConfig(
+  config: Config,
+): {
+  kind: 'legacy' | 'single' | 'multiple';
+  config: NormalizedConfig;
+} {
   if (isLegacyConfig(config)) {
     console.log('config type - "legacy"');
     return {
-      [defaultConfigName]: {
-        name: defaultConfigName,
-        schema: config.schema.path,
-        branch: config.schema.ref,
-        endpoint: config.endpoint,
-        notifications: prioritize<Notifications | false>(
-          config.notifications,
-          notificationsDefault,
-        ),
-        diff: prioritize<Diff | false>(config.diff, diffDefault),
+      kind: 'legacy',
+      config: {
+        [defaultConfigName]: {
+          name: defaultConfigName,
+          schema: config.schema.path,
+          branch: config.schema.ref,
+          endpoint: config.endpoint,
+          notifications: prioritize<Notifications | false>(
+            config.notifications,
+            notificationsDefault,
+          ),
+          diff: prioritize<Diff | false>(config.diff, diffDefault),
+        },
       },
     };
   }
@@ -114,16 +125,19 @@ function normalizeConfig(config: Config): NormalizedConfig {
   if (isSingleEnvironmentConfig(config)) {
     console.log('config type - "single"');
     return {
-      [config.branch]: {
-        name: config.branch,
-        schema: config.schema,
-        branch: config.branch,
-        endpoint: config.endpoint,
-        notifications: prioritize<Notifications | false>(
-          config.notifications,
-          notificationsDefault,
-        ),
-        diff: prioritize<Diff | false>(config.diff, diffDefault),
+      kind: 'single',
+      config: {
+        [config.branch]: {
+          name: config.branch,
+          schema: config.schema,
+          branch: config.branch,
+          endpoint: config.endpoint,
+          notifications: prioritize<Notifications | false>(
+            config.notifications,
+            notificationsDefault,
+          ),
+          diff: prioritize<Diff | false>(config.diff, diffDefault),
+        },
       },
     };
   }
@@ -151,7 +165,10 @@ function normalizeConfig(config: Config): NormalizedConfig {
       }
     }
 
-    return normalized;
+    return {
+      kind: 'multiple',
+      config: normalized,
+    };
   }
 
   throw new Error('Invalid configuration');
@@ -176,12 +193,17 @@ function getGlobalConfig(
 
 export function createConfig(
   rawConfig: Config,
+  setConfigKind: (kind: 'legacy' | 'single' | 'multiple') => void,
   branches: string[] = [],
   fallbackBranch = defaultFallbackBranch,
 ): NormalizedEnvironment {
-  const normalizedConfig = normalizeConfig(rawConfig);
+  const {config: normalizedConfig, kind: configKind} = normalizeConfig(
+    rawConfig,
+  );
 
   let config: NormalizedEnvironment | null = null;
+
+  setConfigKind(configKind);
 
   if (isNormalizedLegacyConfig(normalizedConfig)) {
     config = normalizedConfig[defaultConfigName];
@@ -280,7 +302,7 @@ type Maybe<T> = T | undefined | null;
 type Toggle<T> = T | boolean;
 type Option<T> = Toggle<Maybe<T>>;
 
-// I'm not very proud about it :)
+// I'm not very proud of it :)
 function prioritize<T>(
   child: Option<T>,
   parent: Option<T>,