+++ packages/github/src/helpers/loaders.ts
@@ -62,7 +62,7 @@ export function createFileLoader(config: FileLoaderConfig): FileLoader {
       );
       const {context, repo, owner} = config;
 
-      const result = await context.github.graphql(
+      const result: any = await context.github.graphql(
         createGetFilesQuery(variablesMap),
         {
           repo,
@@ -238,7 +238,14 @@ export async function loadSources({
   ]);
 
   return {
-    old: new Source(oldFile!),
-    new: new Source(newFile!),
+    old: new Source(
+      oldFile!,
+      useEndpoint
+        ? typeof config.endpoint! === 'string'
+          ? config.endpoint
+          : config.endpoint?.url
+        : `${oldPointer.ref}:${oldPointer.path}`,
+    ),
+    new: new Source(newFile!, `${newPointer.ref}:${newPointer.path}`),
   };
 }