+++ packages/github/src/helpers/types.ts
@@ -40,4 +40,5 @@ export interface PullRequest {
   url: string;
   id: number;
   number: number;
+  labels?: Array<{name: string}>;
 }