+++ packages/github/src/schema-diff.ts
@@ -1,5 +1,5 @@
 import * as probot from 'probot';
-import {buildSchema} from 'graphql';
+import { produceSchema } from './helpers/schema';
 import {CheckConclusion, PullRequest} from './helpers/types';
 import {FileLoader, ConfigLoader, loadSources} from './helpers/loaders';
 import {start, complete, annotate} from './helpers/check-runs';
@@ -17,6 +17,7 @@ export async function handleSchemaDiff({
   action,
   context,
   ref,
+  pullRequestNumber,
   repo,
   owner,
   before,
@@ -29,6 +30,7 @@ export async function handleSchemaDiff({
   owner: string;
   repo: string;
   ref: string;
+  pullRequestNumber?: number;
   pullRequests: PullRequest[];
   /***
    * The SHA of the most recent commit on ref before the push
@@ -64,6 +66,7 @@ export async function handleSchemaDiff({
     const branches = pullRequests.map((pr) => pr.base.ref);
     const firstBranch = branches[0];
     const fallbackBranch = firstBranch || before;
+    let isLegacyConfig = false;
 
     logger.info(`fallback branch from Pull Requests: ${firstBranch}`);
     logger.info(`SHA before push: ${before}`);
@@ -71,6 +74,9 @@ export async function handleSchemaDiff({
     // on non-environment related PRs, use a branch from first associated pull request
     const config = createConfig(
       rawConfig as any,
+      (configKind) => {
+        isLegacyConfig = configKind === 'legacy'
+      },
       branches,
       fallbackBranch, // we will probably throw an error when both are not defined
     );
@@ -100,6 +106,17 @@ export async function handleSchemaDiff({
       return;
     }
 
+    if (config.diff.experimental_merge) {
+      if (!pullRequestNumber && pullRequests?.length) {
+        pullRequestNumber = pullRequests[0].number;
+      }
+
+      if (pullRequestNumber) {
+        ref = `refs/pull/${pullRequestNumber}/merge`;
+        logger.info(`[EXPERIMENTAL] Using Pull Request: ${ref}`);
+      }
+    }
+
     const oldPointer: SchemaPointer = {
       path: config.schema,
       ref: config.branch,
@@ -125,14 +142,8 @@ export async function handleSchemaDiff({
     });
 
     const schemas = {
-      old: buildSchema(sources.old, {
-        assumeValid: true,
-        assumeValidSDL: true,
-      }),
-      new: buildSchema(sources.new, {
-        assumeValid: true,
-        assumeValidSDL: true,
-      }),
+      old: produceSchema(sources.old),
+      new: produceSchema(sources.new),
     };
 
     logger.info(`built schemas`);
@@ -152,10 +163,21 @@ export async function handleSchemaDiff({
     logger.info(`changes - ${changes.length}`);
     logger.info(`annotations - ${changes.length}`);
 
-    const summary = createSummary(changes);
+    const summaryLimit = config.diff.summaryLimit || 100;
+
+    const summary = createSummary(changes, summaryLimit, isLegacyConfig);
+
+    const approveLabelName =
+      config.diff.approveLabel || 'approved-breaking-change';
+    const hasApprovedBreakingChangeLabel = pullRequestNumber
+      ? pullRequests[0].labels?.find((label) => label.name === approveLabelName)
+      : false;
 
     // Force Success when failOnBreaking is disabled
-    if (config.diff.failOnBreaking === false) {
+    if (
+      config.diff.failOnBreaking === false ||
+      hasApprovedBreakingChangeLabel
+    ) {
       logger.info('FailOnBreaking disabled. Forcing SUCCESS');
       conclusion = CheckConclusion.Success;
     }
@@ -167,6 +189,9 @@ export async function handleSchemaDiff({
 
     if (config.diff.annotations === false) {
       logger.info(`Anotations are disabled. Skipping annotations...`);
+      annotations = [];
+    } else if (annotations.length > summaryLimit) {
+      logger.info(`Total amount of annotations is over the limit (${annotations.length} > ${summaryLimit}). Skipping annotations...`);
       annotations = [];
     } else {
       logger.info(`Sending annotations (${annotations.length})`);