+++ packages/loaders/loaders/src/index.ts
@@ -10,7 +10,7 @@ import {
   LoadSchemaOptions,
   LoadTypedefsOptions,
 } from '@graphql-tools/load';
-import {GraphQLSchema} from 'graphql';
+import {GraphQLSchema, buildSchema} from 'graphql';
 
 export class LoadersRegistry {
   private loaders: UniversalLoader[] = [];
@@ -35,10 +35,63 @@ export class LoadersRegistry {
   loadSchema(
     pointer: SchemaPointerSingle,
     options: Omit<LoadSchemaOptions, 'loaders'> = {},
+    enableApolloFederation: boolean,
+    enableAWS: boolean,
   ): Promise<GraphQLSchema> {
     return loadSchema(pointer, {
       loaders: this.loaders,
       ...options,
+      ...(enableApolloFederation
+        ? {
+            schemas: [
+              buildSchema(/* GraphQL */ `
+                scalar _FieldSet
+                directive @external on FIELD_DEFINITION
+                directive @requires(fields: _FieldSet!) on FIELD_DEFINITION
+                directive @provides(fields: _FieldSet!) on FIELD_DEFINITION
+                directive @key(fields: _FieldSet!) on OBJECT | INTERFACE
+                directive @extends on OBJECT | INTERFACE
+              `),
+            ],
+          }
+        : {}),
+      ...(enableAWS
+        ? {
+            schemas: [
+              buildSchema(/* GraphQL */ `
+                scalar AWSDate
+                scalar AWSTime
+                scalar AWSDateTime
+                scalar AWSTimestamp
+                scalar AWSEmail
+                scalar AWSJSON
+                scalar AWSURL
+                scalar AWSPhone
+                scalar AWSIPAddress
+                scalar BigInt
+                scalar Double
+
+                directive @aws_subscribe(
+                  mutations: [String!]!
+                ) on FIELD_DEFINITION
+
+                directive @deprecated(
+                  reason: String
+                ) on FIELD_DEFINITION | INPUT_FIELD_DEFINITION | ENUM | ENUM_VALUE
+
+                directive @aws_auth(
+                  cognito_groups: [String!]!
+                ) on FIELD_DEFINITION
+                directive @aws_api_key on FIELD_DEFINITION | OBJECT
+                directive @aws_iam on FIELD_DEFINITION | OBJECT
+                directive @aws_oidc on FIELD_DEFINITION | OBJECT
+                directive @aws_cognito_user_pools(
+                  cognito_groups: [String!]
+                ) on FIELD_DEFINITION | OBJECT
+              `),
+            ],
+          }
+        : {}),
     });
   }